+++ NEWS.md
@@ -11,6 +11,7 @@
 * `prune_vertices` now has an option to avoid creating circles when pruning.
 * Several updates and quality of life improvements for building very large graphs faster.
 * Added the `export()` method, that allows one to export a MetricGraph object as an `sf`, `sp` or `SSN2` object.
+* Added wrappers for `leaflet` and `mapview` as methods.
 
 # MetricGraph 1.3.0
 * Handlers were added in `add_observations()` for situations where observations are projected at the same location, specifically for the `duplicated_strategy` argument.